feat: reorganize domain classes into domain package
