+++ src/main/java/ghostleg/domain/PlayerName.java
@@ -1,4 +1,4 @@
-package ghostleg;
+package ghostleg.domain;
 
 import java.util.regex.Pattern;
