+++ src/test/java/ghostleg/domain/LadderHeightTest.java
@@ -1,4 +1,4 @@
-package ghostleg;
+package ghostleg.domain;
 
 import org.junit.jupiter.api.Test;
