+++ src/test/java/ghostleg/domain/LineTest.java
@@ -1,4 +1,4 @@
-package ghostleg;
+package ghostleg.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
