+++ src/test/java/ghostleg/domain/PointTest.java
@@ -1,4 +1,4 @@
-package ghostleg;
+package ghostleg.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
