+++ src/test/java/ghostleg/domain/PositionTest.java
@@ -1,8 +1,8 @@
-package ghostleg;
+package ghostleg.domain;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class PositionTest {
     @Test
